+++ cmd/fofa/cmd/account.go
@@ -1,0 +1,16 @@
+package cmd
+
+import (
+	"fmt"
+	"github.com/urfave/cli/v2"
+)
+
+// account 子命令
+var accountCmd = &cli.Command{
+	Name:  "account",
+	Usage: "fofa account information",
+	Action: func(ctx *cli.Context) error {
+		fmt.Println(fofaCli.Account)
+		return nil
+	},
+}