+++ cmd/fofa/cmd/cmd.go
@@ -1,13 +1,8 @@
 package cmd
 
 import (
-	"errors"
-	"fmt"
 	"github.com/lubyruffy/gofofa"
-	"github.com/lubyruffy/gofofa/pkg/outformats"
 	"github.com/urfave/cli/v2"
-	"os"
-	"strings"
 )
 
 var (
@@ -15,18 +10,30 @@ var (
 )
 
 var (
-	fofaURL     string // fofa url
-	deductMode  string // deduct Mode
-	fieldString string // fieldString
-	query       string // fieldString
-	size        int    // fetch size
-	outFormat   string // out format
+	fofaURL string // fofa url
 )
 
 // GlobalCommands global commands
 var GlobalCommands = []*cli.Command{
 	searchCmd,
 	accountCmd,
+}
+
+// IsValidCommand valid command name
+func IsValidCommand(cmd string) bool {
+	if len(cmd) == 0 {
+		return false
+	}
+	if cmd[0] == '-' {
+		return false
+	}
+
+	for _, command := range GlobalCommands {
+		if command.Name == cmd {
+			return true
+		}
+	}
+	return false
 }
 
 // GlobalOptions global options
@@ -37,93 +44,6 @@ var GlobalOptions = []cli.Flag{
 		Value:       gofofa.FofaURLFromEnv(),
 		Usage:       "format: <url>/?email=<email>&key=<key>&version=<v2>",
 		Destination: &fofaURL,
-	},
-}
-
-// account 子命令
-var accountCmd = &cli.Command{
-	Name:  "account",
-	Usage: "fofa account information",
-	Action: func(ctx *cli.Context) error {
-		fmt.Println(fofaCli.Account)
-		return nil
-	},
-}
-
-// search 子命令
-var searchCmd = &cli.Command{
-	Name:  "search",
-	Usage: "fofa host search",
-	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:        "query",
-			Aliases:     []string{"q"},
-			Value:       "",
-			Usage:       "fofa query",
-			Destination: &query,
-		},
-		&cli.StringFlag{
-			Name:        "fields",
-			Aliases:     []string{"f"},
-			Value:       "ip,port",
-			Usage:       "visit fofa website for more info",
-			Destination: &fieldString,
-		},
-		&cli.StringFlag{
-			Name:        "outFormat",
-			Aliases:     []string{"o"},
-			Value:       "csv",
-			Usage:       "can be csv/json/xml",
-			Destination: &outFormat,
-		},
-		&cli.IntFlag{
-			Name:        "size",
-			Value:       100,
-			Usage:       "if DeductModeFree set, select free limit size automatically",
-			Destination: &size,
-		},
-		&cli.StringFlag{
-			Name:        "deductMode",
-			Value:       "DeductModeFree",
-			Usage:       "DeductModeFree or DeductModeFCoin",
-			Destination: &deductMode,
-		},
-	},
-	Action: func(ctx *cli.Context) error {
-		// valid same config
-		if len(query) == 0 {
-			return errors.New("fofa query cannot be empty")
-		}
-		fields := strings.Split(fieldString, ",")
-		if len(fields) == 0 {
-			return errors.New("fofa fields cannot be empty")
-		}
-
-		// gen writer
-		outF := os.Stdout
-		var writer outformats.OutWriter
-		switch outFormat {
-		case "csv":
-			writer = outformats.NewCSVWriter(outF)
-		case "json":
-			writer = outformats.NewJSONWriter(outF, fields)
-		case "xml":
-			writer = outformats.NewXMLWriter(outF, fields)
-		default:
-			return fmt.Errorf("unknown outFormat: %s", outFormat)
-		}
-
-		// do search
-		res, err := fofaCli.HostSearch(query, 100, fields)
-		if err != nil {
-			return err
-		}
-
-		// output
-		if err = writer.WriteAll(res); err != nil {
-			return err
-		}
-		return nil
 	},
 }