+++ HISTORY.md
@@ -1,3 +1,8 @@
+
+## v0.0.2
+- support default command to search: ```./fofa port=80```
+- search support -o param to write to file: ```./fofa search -o a.txt port=80```
+
 ## v0.0.1 initial release
 - add search/account subcommand
 - add csv/json/xml output format